Fairly big update: Added recycler view with detail view for users drawings. The serries of events for a user to make a drawing is now My drawings > Add button OR detail view edit button. These buttons take the user to the draw activity.
Room database is now functional - the current object it is storing is
the Drawing object. This has an id, title, maptype, line color and
fragment amount fields.

You can create dummy data for the database on the admin screen. Currently
the drawSettingsActivity does not create an object in the database.